+++ tests/snapshots/columns--with-styles-from.php
@@ -209,7 +209,7 @@ return array (
         'format' => 
         array (
           'format-num-id' => 14,
-          'format-pattern' => 'm/d/yy',
+          'format-pattern' => 'mm-dd-yy',
           'format-category' => 'date',
         ),
         'font' =>