+++ tests/snapshots/columns--with-styles.php
@@ -350,7 +350,7 @@ return array (
         'format' => 
         array (
           'format-num-id' => 14,
-          'format-pattern' => 'm/d/yy',
+          'format-pattern' => 'mm-dd-yy',
           'format-category' => 'date',
         ),
         'font' => 
@@ -386,7 +386,7 @@ return array (
         'format' => 
         array (
           'format-num-id' => 14,
-          'format-pattern' => 'm/d/yy',
+          'format-pattern' => 'mm-dd-yy',
           'format-category' => 'date',
         ),
         'font' => 
@@ -422,7 +422,7 @@ return array (
         'format' => 
         array (
           'format-num-id' => 14,
-          'format-pattern' => 'm/d/yy',
+          'format-pattern' => 'mm-dd-yy',
           'format-category' => 'date',
         ),
         'font' =>